requirements.txt
# SEO Intelligence Agent - Production API Dependencies
# Python 3.10+

fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
beautifulsoup4==4.12.3
pyyaml==6.0.1
pandas==2.2.0
jinja2==3.1.3
python-multipart==0.0.9
python-dotenv==1.0.0
